requirements.txt
# LocalZure Dependencies

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Monitoring and metrics
prometheus-client>=0.19.0

# Configuration
PyYAML>=6.0.1

# HTTP client
httpx>=0.25.0

# Async support
asyncio>=3.4.3

# Persistence (SVC-SB-010)
aiosqlite>=0.19.0

# Docker integration (optional but recommended)
docker>=7.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
